requirements.txt
fastapi
uvicorn[standard]
request